+++ src/main/java/com/drakmyth/minecraft/manufactory/items/RockDrillItem.java
@@ -10,8 +10,8 @@ import java.util.Map;
 import javax.annotation.Nullable;
 
 import com.drakmyth.minecraft.manufactory.LogMarkers;
-import com.drakmyth.minecraft.manufactory.containers.ItemInventory;
-import com.drakmyth.minecraft.manufactory.containers.RockDrillUpgradeContainerProvider;
+import com.drakmyth.minecraft.manufactory.menus.ItemInventory;
+import com.drakmyth.minecraft.manufactory.menus.RockDrillUpgradeContainerProvider;
 import com.drakmyth.minecraft.manufactory.init.ModTags;
 import com.drakmyth.minecraft.manufactory.items.upgrades.IDrillHeadUpgrade;
 import com.drakmyth.minecraft.manufactory.items.upgrades.IMotorUpgrade;
