+++ src/main/java/com/drakmyth/minecraft/manufactory/menus/BallMillContainerProvider.java
@@ -3,7 +3,7 @@
  *  Copyright (c) 2020 Drakmyth. All rights reserved.
  */
 
-package com.drakmyth.minecraft.manufactory.containers;
+package com.drakmyth.minecraft.manufactory.menus;
 
 import com.drakmyth.minecraft.manufactory.LogMarkers;
 
@@ -13,28 +13,28 @@ import org.apache.logging.log4j.Logger;
 import net.minecraft.world.entity.player.Player;
 import net.minecraft.world.entity.player.Inventory;
 import net.minecraft.world.inventory.AbstractContainerMenu;
-import net.minecraft.world.item.ItemStack;
 import net.minecraft.world.MenuProvider;
+import net.minecraft.core.BlockPos;
 import net.minecraft.network.chat.Component;
 import net.minecraft.network.chat.TextComponent;
 import net.minecraftforge.items.wrapper.InvWrapper;
 
-public class RockDrillUpgradeContainerProvider implements MenuProvider {
+public class BallMillContainerProvider implements MenuProvider {
     private static final Logger LOGGER = LogManager.getLogger();
-    private ItemStack stack;
+    private BlockPos pos;
 
-    public RockDrillUpgradeContainerProvider(ItemStack stack) {
-        this.stack = stack;
+    public BallMillContainerProvider(BlockPos pos) {
+        this.pos = pos;
     }
 
     @Override
     public AbstractContainerMenu createMenu(int windowId, Inventory playerInventory, Player player) {
-        LOGGER.debug(LogMarkers.CONTAINER, "Creating Ball Mill upgrade gui...");
-        return new RockDrillUpgradeContainer(windowId, new InvWrapper(playerInventory), player, stack);
+        LOGGER.debug(LogMarkers.CONTAINER, "Creating Ball Mill gui...");
+        return new BallMillContainer(windowId, new InvWrapper(playerInventory), player, pos);
     }
 
     @Override
     public Component getDisplayName() {
-        return new TextComponent("Rock Drill");
+        return new TextComponent("Ball Mill");
     }
 }
