+++ src/main/java/com/drakmyth/minecraft/manufactory/menus/RockDrillUpgradeContainerProvider.java
@@ -3,7 +3,7 @@
  *  Copyright (c) 2020 Drakmyth. All rights reserved.
  */
 
-package com.drakmyth.minecraft.manufactory.containers;
+package com.drakmyth.minecraft.manufactory.menus;
 
 import com.drakmyth.minecraft.manufactory.LogMarkers;
 
@@ -13,28 +13,28 @@ import org.apache.logging.log4j.Logger;
 import net.minecraft.world.entity.player.Player;
 import net.minecraft.world.entity.player.Inventory;
 import net.minecraft.world.inventory.AbstractContainerMenu;
+import net.minecraft.world.item.ItemStack;
 import net.minecraft.world.MenuProvider;
-import net.minecraft.core.BlockPos;
 import net.minecraft.network.chat.Component;
 import net.minecraft.network.chat.TextComponent;
 import net.minecraftforge.items.wrapper.InvWrapper;
 
-public class GrinderContainerProvider implements MenuProvider {
+public class RockDrillUpgradeContainerProvider implements MenuProvider {
     private static final Logger LOGGER = LogManager.getLogger();
-    private BlockPos pos;
+    private ItemStack stack;
 
-    public GrinderContainerProvider(BlockPos pos) {
-        this.pos = pos;
+    public RockDrillUpgradeContainerProvider(ItemStack stack) {
+        this.stack = stack;
     }
 
     @Override
-    public AbstractContainerMenu createMenu(int windowId, Inventory Inventory, Player player) {
-        LOGGER.debug(LogMarkers.CONTAINER, "Creating Grinder gui...");
-        return new GrinderContainer(windowId, new InvWrapper(Inventory), player, pos);
+    public AbstractContainerMenu createMenu(int windowId, Inventory playerInventory, Player player) {
+        LOGGER.debug(LogMarkers.CONTAINER, "Creating Rock Drill upgrade gui...");
+        return new RockDrillUpgradeContainer(windowId, new InvWrapper(playerInventory), player, stack);
     }
 
     @Override
     public Component getDisplayName() {
-        return new TextComponent("Grinder");
+        return new TextComponent("Rock Drill");
     }
 }
